Check both completion conditions when adding ready/failed/do/doError.
Adding a ready after there had been an error would try to push it to
the ready slots, which was null because all these arrays are removed
when an error has occurred. The same was true for do, and the inverse
for doErr and failed after error-free completion.

All these places needed to check both this.completed and this.hasFailed.
Better might be to add Future.prototype.settled() to avoid the need to
repeat the checks.